Move TabTitle to tab-title
